refactor: convert bytes metric from `sum` to `counter`
We need consistency to make sure monitoring is easy to understand.

There is no need to use `sum` metric services for `EdgeBytes`. Return
counters so we can create dashboards easily and keeps consistency with
every other metric in the exporter.

This commit updates metrics description as well, following the GraphQL
API description.