feat: Add support for LoadBalancer Pool health & repo restructuring
Add support for monitoring loadBalancer pool health.

Add a metric to monitor the health of the origin pools.
The repo has been slightly restructured.

Move commonly used structures into a dedicated directory. Rename
packages.

This restructure prepares the project for more info to come and makes
the structure easier to follow and modular.